refactor: melhora de exceção personalizada método 'delete'

+++ dscommerce/src/main/java/com/teste/dscommerce/services/ProductService.java
@@ -1,16 +1,25 @@
 package com.teste.dscommerce.services;
 
+import com.teste.dscommerce.dto.CustomError;
 import com.teste.dscommerce.dto.ProductDTO;
 import com.teste.dscommerce.entities.Product;
+import com.teste.dscommerce.exceptions.DatabaseException;
 import com.teste.dscommerce.exceptions.ResourceNotFoundException;
 import com.teste.dscommerce.mappers.ProductMapper;
 import com.teste.dscommerce.repositories.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 
 
 @Service
@@ -46,24 +55,34 @@ public class ProductService {
 
   @Transactional
   public ProductDTO update(Long id, ProductDTO dto) {
-    Product entity = repository.findById(id)
-        .orElseThrow(() -> new ResourceNotFoundException("Id não encontrado: " + id));
-
-    copyToEntity(dto, entity);
-
-    entity = repository.save(entity);
-    return mapper.toDTO(entity);
+    try {
+      Product entity = repository.getReferenceById(id);
+      copyToEntity(dto, entity);
+      entity = repository.save(entity);
+      return mapper.toDTO(entity);
+    } catch (EntityNotFoundException e) {
+      throw new ResourceNotFoundException("Recurso não encontrado: " + id + "");
+    }
   }
 
-  @Transactional
+  @Transactional(propagation = Propagation.SUPPORTS)
   public void delete(Long id) {
-    repository.deleteById(id);
+    if (!repository.existsById(id)) {
+      throw new ResourceNotFoundException("Recurso não encontrado: " + id);
+    }
+    try {
+      repository.deleteById(id);
+      repository.flush(); // Força a execução SQL imediatamente para capturar exceções do banco dentro do try-catch
+    } catch (DataIntegrityViolationException e) {
+      throw new DatabaseException("Falha de integridade referencial");
+    }
   }
 
-  private void copyToEntity(ProductDTO dto, Product entity) {
-    entity.setName(dto.getName());
-    entity.setDescription(dto.getDescription());
-    entity.setPrice(dto.getPrice());
-    entity.setImgUrl(dto.getImgUrl());
+    private void copyToEntity (ProductDTO dto, Product entity){
+      entity.setName(dto.getName());
+      entity.setDescription(dto.getDescription());
+      entity.setPrice(dto.getPrice());
+      entity.setImgUrl(dto.getImgUrl());
+    }
   }
-}
+
